Updated Chris Williams' fine POD tests to more explicitly treat certain methods as private.  Also added a "strict" mode that doesn't skip over the methods temporarily considered private while we figure out whether they should really be so.  Also removed tabs while I was in there.  Curse you, tabs!
Removed new() methods from classes that are really uninstantiated
mixins.  Calling a nonexistent method is semantically about the same
as calling a constructor that only exists to croak.  Had to remove
tests for the croaks as well.

Documented select_pause_read() and select_resume_read().

Added leading underscores to a bunch of helper functions and methods
that should really have been private.  There are more to come, but
their privacy status is less obvious.