Disable an assertion that system() must return 0 during $SIG{CHLD}.
Perl has a known issue where system() can return -1 (failure) when a
$SIG{CHLD} handler is in effect, even if it succeeds.  It's not POE's
it succeeds when a $SIG{CHLD} handler is in effect.  It's not POE's
duty to work around the problem.  Proper, reliable work-arounds would
incur global overhead.  Use {local $SIG{CHLD}='DEFAULT'; system(...)}
as needed instead.  And watch the skies for a Perl fix:
https://rt.perl.org/rt3/Ticket/Display.html?id=105700