Fix POE::Wheel::FollowTail test hang in wheel_tail.pm
Windows locks files and directories while they're open.  It's not
possible to unlink() them in this state.

The file reset test tried to unlink() a test log to reset it.  The
unlink() was silently failing, and the test program was waiting
indefinitely for the reset.

Added a SKIP section around those tests so they don't run on Windows.
Fixes welcome, because I'd really like Windows users to be able to
detect file resets.

Replaced the "skip_all" plan when networking isn't available with a
SKIP section around the network-based tests.  Filesystem-based tests
will still run when networking isn't available... unless the tests are
running on Windows.